[TASK] Move some templating tests to functionals
For TYPO3 12LTS, we need functional tests for theses tests in order
to resolved loaded extensions.
