+++ Tests/Unit/Templating/TemplateRegistryTest.php
@@ -74,39 +74,4 @@ final class TemplateRegistryTest extends UnitTestCase
             TemplateRegistry::get('')
         );
     }
-
-    /**
-     * @test
-     */
-    public function getForExistingTemplateFileNameReturnsTemplate(): void
-    {
-        self::assertInstanceOf(
-            Template::class,
-            TemplateRegistry::get('EXT:oelib/Tests/Functional/Fixtures/Template.html')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getForExistingTemplateFileNameCalledTwoTimesReturnsNewInstance(): void
-    {
-        self::assertNotSame(
-            TemplateRegistry::get('EXT:oelib/Tests/Functional/Fixtures/Template.html'),
-            TemplateRegistry::get('EXT:oelib/Tests/Functional/Fixtures/Template.html')
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getForExistingTemplateFileNameReturnsProcessedTemplate(): void
-    {
-        $template = TemplateRegistry::get('EXT:oelib/Tests/Functional/Fixtures/Template.html');
-
-        self::assertSame(
-            "Hello world!\n",
-            $template->getSubpart()
-        );
-    }
 }
